fix: few stuff
- add spaces at selected item counter at the top left corner, should be "1 / 2" instead of "1/2"
- clicking on the selected item counter should trigger Select all, then Deselect all
- text size of menu items Copy number to clipboard and Delete are too small
- clicking on 3 dots, the floating ones on the right side should have circle ripple, not square
- fix navigation bar issue add todo because it's a Google's problem

Left to check:
- ripple color test implemented for selection
- light mode icon color messed up

+++ commons/src/main/kotlin/com/simplemobiletools/commons/compose/extensions/DragHandler.kt
@@ -27,7 +27,7 @@ fun Modifier.dragHandler(
     detectDragGesturesAfterLongPress(
         onDragStart = { offset ->
             lazyListState.gridItemKeyAtPosition(offset)?.let { key ->
-                if (!selectedIds.value.contains(key.toLong())) {
+                if (!selectedIds.value.contains(key)) {
                     haptics.performHapticFeedback(HapticFeedbackType.LongPress)
                     initialKey = key
                     currentKey = key
